Grouping test files and their associated classes together
