+++ bigbluebutton-tests/puppeteer/chat/chat.obj.js
@@ -1,12 +1,12 @@
-const Page = require('./core/page');
-const Send = require('./chat/send');
-const Clear = require('./chat/clear');
-const Copy = require('./chat/copy');
-const Save = require('./chat/save');
-const Poll = require('./chat/poll');
-const MultiUsers = require('./user/multiusers');
+const Page = require('../core/page');
+const Send = require('./send');
+const Clear = require('./clear');
+const Copy = require('./copy');
+const Save = require('./save');
+const Poll = require('./poll');
+const MultiUsers = require('../user/multiusers');
 const { toMatchImageSnapshot } = require('jest-image-snapshot');
-const { MAX_CHAT_TEST_TIMEOUT } = require('./core/constants'); // core constants (Timeouts vars imported)
+const { MAX_CHAT_TEST_TIMEOUT } = require('../core/constants'); // core constants (Timeouts vars imported)
 
 expect.extend({ toMatchImageSnapshot });
