+++ bigbluebutton-tests/puppeteer/events/trigger.test.js
