+++ bigbluebutton-tests/puppeteer/presentation/presentation.obj.js
@@ -1,7 +1,7 @@
-const Page = require('./core/page');
-const Presentation = require('./presentation/presentation');
+const Page = require('../core/page');
+const Presentation = require('./presentation');
 const { toMatchImageSnapshot } = require('jest-image-snapshot');
-const { MAX_PRESENTATION_TEST_TIMEOUT } = require('./core/constants'); // core constants (Timeouts vars imported)
+const { MAX_PRESENTATION_TEST_TIMEOUT } = require('../core/constants'); // core constants (Timeouts vars imported)
 
 expect.extend({ toMatchImageSnapshot });
