+++ bigbluebutton-tests/puppeteer/virtualizedlist/virtualizedlist.obj.js
@@ -1,5 +1,5 @@
-const VirtualizedList = require('./virtualizedlist/virtualize');
-const { TEST_DURATION_TIME } = require('./core/constants');
+const VirtualizedList = require('./virtualize');
+const { TEST_DURATION_TIME } = require('../core/constants');
 const { toMatchImageSnapshot } = require('jest-image-snapshot');
 
 expect.extend({ toMatchImageSnapshot });
