+++ bigbluebutton-tests/puppeteer/webcam/webcam.test.js
