+++ bigbluebutton-tests/puppeteer/webcam/webcamlayout.obj.js
@@ -1,7 +1,7 @@
-const Page = require('./core/page');
-const Share = require('./webcam/share');
+const Page = require('../core/page');
+const Share = require('./share');
 const { toMatchImageSnapshot } = require('jest-image-snapshot');
-const { MAX_WEBCAM_LAYOUT_TEST_TIMEOUT } = require('./core/constants');
+const { MAX_WEBCAM_LAYOUT_TEST_TIMEOUT } = require('../core/constants');
 
 expect.extend({ toMatchImageSnapshot });
