+++ bigbluebutton-tests/puppeteer/webcam/webcamlayout.test.js
