+++ bigbluebutton-tests/puppeteer/whiteboard/whiteboard.test.js
